refactor: Moved nodes to new module
